Expose a more direct lightweight serialization mechanism
Introduce a mechanism to convert Mphf into a series of slices and
MphfRef which can take a series of slices as input and access it without
copying the underlying data. This is useful when serializing to disk and
then accessing via mmap but does have caveats that it's restricted to
little-endian systems and that the data is stored on 64-bit aligned
addresses within a file.